refactor: split AccountController into Admin, Client, ServiceClient, and Prestataire controllers; update all views and layout for new controller structure and routing
