+++ Views/Account/RegisterAdmin.cshtml
@@ -143,7 +143,7 @@
             </div>
         }
 
-        <form asp-action="RegisterAdmin" asp-controller="Account" method="post">
+        <form asp-action="RegisterAdmin" asp-controller="Admin" method="post">
             @Html.AntiForgeryToken()
             <input type="hidden" name="role" value="Admin" />
