+++ Views/Admin/AdminDashboard.cshtml
@@ -77,7 +77,7 @@
                 </td>
                 <td>
                     @if (d.Etat == "non valide") {
-                        <form asp-action="FinalRejectDemandePrestataire" method="post" style="display:inline" onsubmit="window.open('https://mail.google.com/mail/?view=cm&fs=1&to=@d.Email&su=Your demand has been rejected&body=Reason: @System.Net.WebUtility.UrlEncode(d.Reason ?? "")', '_blank');">
+                        <form asp-action="FinalRejectDemandePrestataire" asp-controller="Admin" method="post" style="display:inline" onsubmit="window.open('https://mail.google.com/mail/?view=cm&fs=1&to=@d.Email&su=Your demand has been rejected&body=Reason: @System.Net.WebUtility.UrlEncode(d.Reason ?? "")', '_blank');">
                             <input type="hidden" name="id" value="@d.Iddemandeprestataire" />
                             <button type="submit" class="btn btn-danger btn-sm">Non Valide Final</button>
                         </form>
@@ -87,7 +87,7 @@
                         <div class="modal fade" id="createAccountModal_@d.Iddemandeprestataire" tabindex="-1" aria-labelledby="createAccountModalLabel_@d.Iddemandeprestataire" aria-hidden="true">
                           <div class="modal-dialog">
                             <div class="modal-content">
-                              <form asp-action="CreatePrestataireAccount" method="post">
+                              <form asp-action="CreatePrestataireAccount" asp-controller="Admin" method="post">
                                 <div class="modal-header">
                                   <h5 class="modal-title" id="createAccountModalLabel_@d.Iddemandeprestataire">Create Prestataire Account</h5>
                                   <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
