+++ Views/Home/Index.cshtml
@@ -503,7 +503,7 @@
                     <div class="role-content">
                         <h4>I'm a Client</h4>
                         <p>Looking to hire top talent? Find skilled professionals for any project, big or small.</p>
-                        <a asp-controller="Account" asp-action="RegisterClient" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Client</a>
+                        <a asp-controller="Client" asp-action="RegisterClient" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Client</a>
                     </div>
                 </div>
             </div>
@@ -519,7 +519,7 @@
                     <div class="role-content">
                         <h4>I'm a Service Client</h4>
                         <p>Need specialized or ongoing services? Connect with providers offering tailored solutions.</p>
-                        <a asp-controller="Account" asp-action="RegisterServiceClient" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Service Client</a>
+                        <a asp-controller="ServiceClient" asp-action="RegisterServiceClient" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Service Client</a>
                     </div>
                 </div>
             </div>
@@ -535,7 +535,7 @@
                     <div class="role-content">
                         <h4>I'm a Prestataire</h4>
                         <p>Offer your expertise, find exciting projects, and expand your professional reach.</p>
-                        <a asp-controller="Account" asp-action="RegisterPrestataire" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Prestataire</a>
+                        <a asp-controller="Prestataire" asp-action="RegisterPrestataireDemande" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Prestataire</a>
                     </div>
                 </div>
             </div>
@@ -551,7 +551,7 @@
                     <div class="role-content">
                         <h4>I'm an Admin</h4>
                         <p>Oversee platform operations, support users, and ensure Hirfa runs smoothly.</p>
-                        <a asp-controller="Account" asp-action="RegisterAdmin" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Admin</a>
+                        <a asp-controller="Admin" asp-action="RegisterAdmin" class="btn btn-modern btn-modern-primary mt-auto">Sign Up as Admin</a>
                     </div>
                 </div>
             </div>
