+++ Views/Prestataire/PrestataireDemands.cshtml
@@ -195,7 +195,7 @@
                                 </span>
                             </td>
                             <td class="text-center">
-                                <a asp-action="PrestataireDemandDetails" asp-route-id="@d.Iddemandeprestataire" class="btn btn-action btn-outline-primary" title="View Details">
+                                <a asp-action="PrestataireDemandDetails" asp-controller="Prestataire" asp-route-id="@d.Iddemandeprestataire" class="btn btn-action btn-outline-primary" title="View Details">
                                     <i class="bi bi-eye-fill"></i> View
                                 </a>
                                 @* Add other actions like Edit/Delete here if needed, with appropriate auth checks *@
