+++ Views/ServiceClient/ServiceClientDashboard.cshtml
@@ -27,11 +27,11 @@
                     <td>
                         @if (d.Etat == Hirfa.Web.Models.DemandeclientStatus.Pending)
                         {
-                            <form asp-action="AcceptDemandeClient" method="post" style="display:inline">
+                            <form asp-action="AcceptDemandeClient" asp-controller="ServiceClient" method="post" style="display:inline">
                                 <input type="hidden" name="id" value="@d.Iddemandeclient" />
                                 <button type="submit" class="btn btn-success btn-sm">Accept</button>
                             </form>
-                            <form asp-action="RejectDemandeClient" method="post" style="display:inline">
+                            <form asp-action="RejectDemandeClient" asp-controller="ServiceClient" method="post" style="display:inline">
                                 <input type="hidden" name="id" value="@d.Iddemandeclient" />
                                 <button type="submit" class="btn btn-danger btn-sm">Reject</button>
                             </form>
